requirements.txt
openai
langchain
python-dotenv
